Extract shared text-collation logic into resolver
SQLContainsTextCondition and SQLEndsWithCondition carried a verbatim
copy of the same collation-resolution logic (Result and Identifiable
resolve overloads, the base-identifier scan-path fast path with its
row-invariant name cache, the schema-class collation read, and the
any()/all() per-property path). Duplicated logic drifts: a fix to one
node's collation handling would silently miss the other.

Move it into a single package-private TextCollationResolver. Each node
holds its own instance (private final) because the resolver caches the
property name derived from that node's left operand, and the two nodes
carry different operands, so the cache must never be shared. Operands
are passed per call, not snapshotted, because the parser and
MatchWhereBuilder set left/right after construction. The nodes keep
their own inline comparisons (endsWith / indexOf), applying collation
through the shared apply() helper. Behavior is unchanged.
